% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wellmap_long.R
\name{make_wellmap_long}
\alias{make_wellmap_long}
\title{Launch an interactive webpage to convert well maps to long format}
\usage{
make_wellmap_long()
}
\description{
There are no arguments for this function, as it launches a Shiny app for user input. You can either input values directly by double clicking on cells, or paste data
from Excel or Google Sheets. If pasting from Excel/Google Sheets, only copy well values, not headers (column numbers) or row names (letters like A, B, etc.)
}
\details{
There are buttons on the bottom to download long-converted data to .csv or .txt (tab delimited) files.
}
\examples{
make_wellmap_long() # Launch shiny app

}
