% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_Growth_Calculations.R
\name{double_times}
\alias{double_times}
\title{Calculate doubling time in exponential phase}
\usage{
double_times(df, time, measure, c1, c2)
}
\description{
Calculate doubling time in exponential phase
}
